requirements.txt
git+https://github.com/Rapptz/discord.py
dnspython==1.16.0
PyNaCl==1.3.0
asyncio
requests
async-timeout
youtube_dl>=2020.05.29
discord.py[voice]==1.2.5
colorlog
websockets>=6.0,<7.0
chardet
opuslib